% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDP.R
\docType{data}
\name{GDP}
\alias{GDP}
\title{World Bank GDP data}
\format{
A data frame with columns:
\describe{
 \item{country_name}{Full country names.}
 \item{year}{Year of measurement.}
 \item{GDP}{Gross Demostic Product.}
}
}
\source{
World Bank
<https://data.worldbank.org/indicator/NY.GDP.MKTP.CD>, downloaded 2021-10-15
}
\usage{
GDP
}
\description{
This dataset contains GDP data of the countries in the world since 1960.
}
\keyword{datasets}
